Move sound ROMs to SDRAM
There isn't enough memory bandwith with the current SDRAM controller to MUX all of the ROMs without graphics/audio glitches.

Thankfully the game doesn't crash, because the main and sound CPU ROMs are prioritised over the other.

If we could squeeze a little more performance out of the SDRAM controller then the game would run glitch-free.