Fix tilemap layer pixel delay
The tilemap layers are delaying their pixel data because they latch the tile row too late.

For example, if the tile row is latched on the last pixel (offset=7) of a tile, then the palette won't latch the colour data until the first pixel (offset=0) of the next tile. This means the actual pixel data won't be visible until the second pixel (offset=1) of the next tile, which is too late.

By latching tile data earlier and selecting the next pixel ahead, we can ensure that the pixels will be visible at the correct time.