Fix sprite code masking
The lower 8 bits of the sprite code need to be masked, depending on the size of the sprite.

The masks look something like this:

8x8:   1111
16x16: 1110
32x32: 1100
64x64: 1000

This is because, for example, a 16x16 sprite can't be offset by a 8 pixels (i.e. it can only be offset on 16 pixel boundaries).

Most games don't leave the lower bits set for larger sprites, but this was definitely a problem with the helicopter sprite in Silkworm.